replace each given character to other e.g. blank with %20
replace each given character to other e.g. blank with %20? (solution)
Write a Java program to replace a given character in a String to other provided character, for example if you are asked to replace each blank in a String with %20, similar to URL encoding done by the browser, so that Server can read all request parameters. For example if the input is "Java is Great" and asked to replace space with %20 then it should be "Java%20is%20Great".