Implement a custom Threadpool in java
Implementation of thread pool , using LinkedBlockingQueue for tasks to do, and an array for n number of threads in pull.
this is little advance version of producer consumer scenario. here threads in thread pool are consumers, these threads keeps on picking tasks from queue (LinkedBlockingQueue), if queue is empty, thread goes to waiting state on queue till producer of task(main method in this example) calls notify after adding a new task on the queue.